Added review functionality

+++ src/handlers/confirmation.js
@@ -4,6 +4,8 @@
  * Display the confirmation page
  */
 const BaseHandler = require('./base')
+const SubmissionsApi = require('../api/submissions')
+const submissionsApi = new SubmissionsApi()
 
 module.exports = class ConfirmHandler extends BaseHandler {
   constructor (...args) {
@@ -18,6 +20,14 @@ module.exports = class ConfirmHandler extends BaseHandler {
    * @returns {Promise<*>}
    */
   async doGet (request, h) {
+    const cache = await request.cache().get()
+    const submission = await submissionsApi.getById(cache.submissionId)
+
+    // If the submission status is not submitted, throw an error
+    if (submission.status !== 'SUBMITTED') {
+      throw new Error('Illegal access of the confirmation page')
+    }
+
     await request.cache().drop()
     request.cookieAuth.clear()
     return h.view(this.path)
